+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.4.0] - 2022-12-10
+### Added
+- Add links
+
 ## [1.3.1] - 2022-12-07
 ### Added
 - Suppress nullable warning in generated source.